Fase 3

+++ entity/src/main/java/com/gestor/seguimiento/PlanSeccionPK.java
@@ -19,7 +19,7 @@ public class PlanSeccionPK implements Serializable {
 
     @Basic(optional = false)
     @Column(name = "codigo_establecimiento")
-    private short codigoEstablecimiento;
+    private int codigoEstablecimiento;
     @Basic(optional = false)
     @Column(name = "cod_titulo")
     private int codTitulo;
@@ -30,17 +30,17 @@ public class PlanSeccionPK implements Serializable {
     public PlanSeccionPK() {
     }
 
-    public PlanSeccionPK(short codigoEstablecimiento, int codTitulo, int codSeccion) {
+    public PlanSeccionPK(int codigoEstablecimiento, int codTitulo, int codSeccion) {
         this.codigoEstablecimiento = codigoEstablecimiento;
         this.codTitulo = codTitulo;
         this.codSeccion = codSeccion;
     }
 
-    public short getCodigoEstablecimiento() {
+    public int getCodigoEstablecimiento() {
         return codigoEstablecimiento;
     }
 
-    public void setCodigoEstablecimiento(short codigoEstablecimiento) {
+    public void setCodigoEstablecimiento(int codigoEstablecimiento) {
         this.codigoEstablecimiento = codigoEstablecimiento;
     }
